requirements.txt
pandas
numpy
streamlit
stremlit_chat
beautifulsoup4
googlesearch-python